Adjust expected test data for a single test issue
There were two differences between the `Original.php` and
`Expected.php` files:

* Setting of the default to function parameters
* Change for the concatintion assignment in the array

In order to simplify the testing and build custom rules one-by-one,
the `Expected.php` was adjusted for now to have a single change.

This can be revereted when at least one of the two custom rules
will be developed.